add: LLM model selection function
Added a frontend button wo select the desired model from a list of available
models. The list is dynamically setup by getting the current list from the
/get-models endpoint on app startup. This endpoint imports the list from the
llm/llm_api.py file with the models_to_json() funtion. Selecting a model
sends a post request to the /set-model endpoint that saves the current model
to give it to the desk_query() (formerly main) function. In the main function
the selected model is evaluated and if a new model or client is needed it is
created using the llm_api's get_model_client() function.

ATTENTION: The application of the created llm client is currently not tested.